perf(decoder): shrink FLACDecoder ~100 bytes (int16_t coefs + lazy metadata) (#8)
I went spelunking looking for zero-cpu-cost memory wins and found a
couple. Nothing earth shattering, but I figured every byte counts. Did a
ton of benchmarking on a real esp32-s3 device to ensure I wasn't mucking
anything up, but as always, please review with a fine-toothed comb. 😅

The actual git commit message:

Narrow LpcState::coefs from int32_t[32] to int16_t[32] (64 B saved):
FLAC qlp coefficient precision is bounded to <=15 bits per RFC 9639
§9.2.6, so values fit in int16_t with no loss. Updates the Xtensa LPC
assembly to load coefs via l16si (halved offsets, addx4->addx2 in the
generic path), the C++ fallback signatures, FIXED_COEFFICIENTS tables,
and the test_lpc_asm reference vectors + generator.

Defer max_metadata_sizes_ (32 B inline -> 4 B pointer) and
metadata_blocks_ to heap allocation; both are lazily populated only when
the opt-in set_max_metadata_size() / metadata-storage features are used.
Repack HeaderParseState (narrow type to uint8_t, drop redundant/unused
data_capacity). Also adds an in-loop bound check on the wasted-bits
shift so pathological input is rejected immediately. Hardware-verified
on ESP32-S3: zero perf regression vs baseline (917.6 ms full-frame,
32.7x real-time), 192/192 LPC assembly test vectors pass.

+++ src/xtensa/lpc_xtensa.h
@@ -62,7 +62,8 @@ extern "C" {
 /// @param buffer Pointer to sub_frame_buffer (int32_t*) - contains warm-up samples followed by
 /// residuals
 /// @param num_samples Total number of samples in the buffer
-/// @param coefficients Pointer to LPC coefficients array (int32_t*)
+/// @param coefficients Pointer to LPC coefficients array (int16_t*; loaded via l16si which
+///                     sign-extends transparently into a 32-bit register for mull)
 /// @param order Number of coefficients (predictor order)
 /// @param shift Shift amount for the prediction
 ///
@@ -70,7 +71,7 @@ extern "C" {
 ///       restored sample values.
 /// @note Optimized for orders 1-12 with fully unrolled loops; uses a generic
 ///       loop for higher orders.
-void restore_lpc_32bit_asm(int32_t* buffer, size_t num_samples, const int32_t* coefficients,
+void restore_lpc_32bit_asm(int32_t* buffer, size_t num_samples, const int16_t* coefficients,
                            uint32_t order, int32_t shift);
 
 /// @brief Optimized 64-bit assembly implementation of linear prediction restoration for FLAC
@@ -91,7 +92,7 @@ void restore_lpc_32bit_asm(int32_t* buffer, size_t num_samples, const int32_t* c
 ///       restored sample values.
 /// @note Optimized for orders 1-12 with fully unrolled loops; uses a generic
 ///       loop for higher orders.
-void restore_lpc_64bit_asm(int32_t* buffer, size_t num_samples, const int32_t* coefficients,
+void restore_lpc_64bit_asm(int32_t* buffer, size_t num_samples, const int16_t* coefficients,
                            uint32_t order, int32_t shift);
 
 #endif  // FLAC_LPC_XTENSA_ENABLED
